base: omnistyle: Stinky workaround for header shadow and tinted images
Issue: Some of our headers uses textures that are tinted with android:tint, but also the headers are tinted with PorterDuff method.
This situation makes the system accent goes full black when even a 1% of 'shadow' and removes all the tint from the headers (making them black, and looking fugly on dark themes)

So, as a bad alternative method, let's use a new view based on the qs background but fully black to work the same as the previous shadow, making the tint issue gone.